detector/requirements.txt
kafka-python==2.0.2
influxdb==5.3.1